src/main/java/com/nicolasf/qualibrate/controller/ProjectController.java
package com.nicolasf.qualibrate.controller;

import com.nicolasf.qualibrate.domain.Project;
import com.nicolasf.qualibrate.service.ProjectService;
import java.util.List;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/v1/api/projects")
public class ProjectController {

    private final ProjectService projectService;

    public ProjectController(ProjectService projectService) {
        this.projectService = projectService;
    }

    @GetMapping
    public List<Project> findAll() {
        return projectService.findAll();
    }

    @PostMapping
    public Project create(@RequestBody Project project) {
        return projectService.create(project);
    }
}
